cog refactor 2
