+++ src/beefcmd/cogs/incantations_cog.py
@@ -1,5 +1,6 @@
 import discord
 from discord.ext import commands
+import beefcmd.incantations.vicious_mockery
 
 class IncantationsCog(commands.Cog):
     def __init__(self, bot):
@@ -7,7 +8,7 @@ class IncantationsCog(commands.Cog):
         
     @discord.app_commands.command(name="vicious_mockery", description="cast vicious mockery on someone")
     async def vicious_mockery(self, interaction: discord.Interaction, victim: discord.Member):
-        print("fuck")
+        await beefcmd.incantations.vicious_mockery.insult(interaction, victim)
 
 async def setup(bot):
     print("incantation cog setup")
