+++ src/responses.py
@@ -1,6 +1,5 @@
 import json
 import os
-import random
 import discord
 from json_handling import load_element
 
@@ -20,20 +19,3 @@ async def get_response(message: discord.Message):
             else:
                 await message.reply(content)
                 return
-        
-    
-async def get_joke_response_positive(member: discord.Member):     
-    jokertag = type('Joker', (object,), {"mention": member.mention})()
-    responses = load_element("responses.json", "joke_responses_positive")
-        
-    chosen_response = random.choice(responses)
-    chosen_response = chosen_response.format(joker=jokertag)
-    return chosen_response
-
-async def get_joke_response_negative( member: discord.Member):     
-    jokertag = type('Joker', (object,), {"mention": member.mention})()
-    responses = load_element("responses.json", "joke_responses_negative")
-        
-    chosen_response = random.choice(responses)
-    chosen_response = chosen_response.format(joker=jokertag)
-    return chosen_response
